+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+  - User, session and configuration OIDC support fixed predefined redirect-url 
   - Header authentication extended to support active-claim
 
 ## [5.11.1] - 2021-03-22